+++ examples/nextjs/pages/index.js
@@ -1,44 +1,17 @@
-import { useEffect, useState } from 'react';
 import Head from 'next/head'
 import Image from 'next/image'
 import { Convo } from "@theconvospace/sdk"
-import { ethers } from "ethers";
 import styles from '../styles/Home.module.css'
+import { ConnectButton } from '@rainbow-me/rainbowkit';
+import { useAccount, useNetwork, useSigner  } from 'wagmi';
+import NextLink from 'next/link';
 
 export default function Home() {
 
   let ConvoInstance = new Convo('CSCpPwHnkB3niBJiUjy92YGP6xVkVZbWfK8xriDO');
-  let [web3, setWeb3] = useState(undefined);
-  let [accounts, setAccounts] = useState(undefined);
-
-  useEffect(() => {
-    authWeb3();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  async function authWeb3() {
-
-    if (Boolean(window.ethereum) == true) {
-
-      ethereum.autoRefreshOnNetworkChange = false;
-      let w3 = new ethers.providers.Web3Provider(window.ethereum);
-      setWeb3(w3);
-      ethereum.request({ method: 'eth_requestAccounts' }).then(setAccounts);
-
-    }
-    else if (window.web3) {
-      setWeb3(web3)
-      web3.currentProvider.enable().then(setAccounts);
-    }
-    else {
-      console.log('Get Web3.')
-    }
-  }
-
-  async function clearAuth() {
-    setWeb3(undefined);
-    setAccounts(undefined);
-  }
+  const { address } = useAccount();
+  const { chain } = useNetwork();
+  const { data: signer } = useSigner();
 
   return (
     <div className={styles.container}>
@@ -58,16 +31,7 @@ export default function Home() {
         </p>
 
         <p className={styles.description}>
-          {
-            accounts === undefined ? (
-              <button onClick={authWeb3}>Authenticate</button>
-            ) : (
-              <>
-                {accounts[0]}
-                &nbsp;<button onClick={clearAuth}>X</button>
-              </>
-            )
-          }
+          <ConnectButton />
         </p>
 
         <div className={styles.grid}>
@@ -103,14 +67,11 @@ export default function Home() {
 
           <div onClick={async () => {
 
-            let timestamp = Date.now();
-            let data = ConvoInstance.auth.getSignatureData(accounts[0], timestamp);
-            let signature = await web3.send(
-              'personal_sign',
-              [ethers.utils.hexlify(ethers.utils.toUtf8Bytes(data)), accounts[0].toLowerCase()]
-            );
-            ConvoInstance.auth.authenticate(accounts[0], signature, timestamp, "ethereum").then(console.log);
-
+            let data = ConvoInstance.auth.getSignatureDataV2('https://convosdk-examples-nextjs.vercel.app', address, chain.id);
+            signer.signMessage(data).then(async (sig) => {
+              const token = await ConvoInstance.auth.authenticateV2(data, sig);
+              console.log('token', token);
+            });
           }} className={styles.card}>
             <h2>Authenticate &rarr;</h2>
             <p>Get an Authentication Token using the SDK.</p>
@@ -119,7 +80,7 @@ export default function Home() {
           <div onClick={() => {
 
             ConvoInstance.auth.validate(
-              accounts[0],
+              address,
               ""
             ).then(console.log);
 
@@ -130,7 +91,7 @@ export default function Home() {
 
           <div onClick={() => {
             let timestamp = Date.now();
-            console.log(ConvoInstance.auth.getSignatureData(accounts[0], timestamp));
+            console.log(ConvoInstance.auth.getSignatureData(address, timestamp));
           }} className={styles.card}>
             <h2>Get Signature Data &rarr;</h2>
             <p>Get the Data used for signing and authentication.</p>
@@ -138,10 +99,9 @@ export default function Home() {
 
           <div onClick={() => {
             console.log(ConvoInstance.auth.getSignatureDataV2(
-              'theconvo.space',
-              'https://theconvo.space/',
-              accounts[0],
-              '1')
+              'https://convosdk-examples-nextjs.vercel.app',
+              address,
+              chain.id)
             );
           }} className={styles.card}>
             <h2>Signature Data V2 &rarr;</h2>
@@ -166,7 +126,7 @@ export default function Home() {
 
           <div onClick={() => {
             ConvoInstance.comments.create(
-              accounts[0],
+              address,
               "",
               "New",
               "KIGZUnR4RzXDFheXoOwo",
@@ -179,7 +139,7 @@ export default function Home() {
 
           <div onClick={() => {
             ConvoInstance.comments.delete(
-              accounts[0],
+              address,
               "",
               "01fh3enc55ppnh7g7r4sz9wvp5"
             ).then(console.log)
@@ -219,16 +179,44 @@ export default function Home() {
         <div className={styles.grid}>
 
           <div onClick={() => {
-            ConvoInstance.omnid.getTrustScore('0xa28992A6744e36f398DFe1b9407474e1D7A3066b').then(console.log);
+            ConvoInstance.omnid.getTrustScore('0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045').then(console.log);
           }} className={styles.card}>
             <h2>Get Identity Data&rarr;</h2>
             <p>Get data for an Ethereum address using the SDK.</p>
           </div>
 
         </div>
+
+        <p className={styles.description}>
+          Components
+        </p>
+
+        <div className={styles.grid}>
+
+          <NextLink href='/embeds'>
+            <div className={styles.card}>
+              <h2>Embeds</h2>
+              <p>Get data for an Ethereum address using the SDK.</p>
+            </div>
+          </NextLink>
+
+          <NextLink href='/comment'>
+            <div className={styles.card}>
+              <h2>Comment</h2>
+              <p>A React Component to showcase a single comment.</p>
+            </div>
+          </NextLink>
+
+          <NextLink href='/commentSection'>
+            <div className={styles.card}>
+              <h2>Comment Section</h2>
+              <p>A React Component for a simple comment section.</p>
+            </div>
+          </NextLink>
+
+
+        </div>
       </main>
-
-
 
       <footer className={styles.footer}>
         <a